visa_la_vista/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
cryptography==46.0.7
Django==5.2.13
django-filter==25.2
djangorestframework==3.17.1
dnspython==2.8.0
drf-nested-routers==0.95.0
drf-yasg==1.21.15
email-validator==2.3.0
fastapi==0.136.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.8
packaging==26.0
pillow==12.2.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyMySQL==1.1.2
pytz==2026.1.post1
PyYAML==6.0.3
SQLAlchemy==2.0.49
sqlmodel==0.0.38
sqlparse==0.5.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.46.0
python-dotenv==1.2.2
boto3==1.42.97
botocore==1.42.97
django-storages==1.14.6
jmespath==1.1.0
python-dateutil==2.9.0
s3transfer==0.16.1
six==1.17.0